+++ src/hyp3_mintpy/__main__.py
@@ -10,6 +10,7 @@ from hyp3lib.aws import upload_file_to_s3
 from hyp3lib.fetch import write_credentials_to_netrc_file
 
 from hyp3_mintpy.process import process_mintpy
+from hyp3_mintpy.util import upload_file_to_s3_with_publish_access_keys
 
 
 def main() -> None:
@@ -20,14 +21,27 @@ def main() -> None:
 
     # TODO: Your arguments here
     parser.add_argument('--job-name', help='The name of the HyP3 job', required=False)
-    parser.add_argument(
-        '--prefix', help='Folder that contains multiburst products in the volcsarvatory bucket', required=False
-    )
+    parser.add_argument('--input-bucket', help='Bucket with multiburst products', required=False)
+    parser.add_argument('--input-prefix', help='Prefix that contains multiburst products', required=False)
+
+    parser.add_argument('--start-date', type=str, help='Start date for the timeseries (YYYY-MM-DD)')
+    parser.add_argument('--end-date', type=str, help='End date for the timeseries (YYYY-MM-DD)')
     parser.add_argument(
         '--min-coherence', default=0.01, type=float, help='The minimum coherence to process', required=False
     )
-    parser.add_argument('--start-date', type=str, help='Start date for the timeseries (YYYY-MM-DD)')
-    parser.add_argument('--end-date', type=str, help='End date for the timeseries (YYYY-MM-DD)')
+    parser.add_argument(
+        '--publish-bucket',
+        type=str,
+        default=None,
+        help='Additionally, publish products to this bucket. Necessary credentials must be provided '
+        'via the `PUBLISH_ACCESS_KEY_ID` and `PUBLISH_SECRET_ACCESS_KEY` environment variables.',
+    )
+    parser.add_argument(
+        '--publish-prefix',
+        type=str,
+        default=None,
+        help='Prefix for the bucket where the products will be published',
+    )
 
     args = parser.parse_args()
 
@@ -48,7 +62,8 @@ def main() -> None:
 
     product_file = process_mintpy(
         job_name=args.job_name,
-        prefix=args.prefix,
+        input_bucket=args.input_bucket,
+        input_prefix=args.input_prefix,
         min_coherence=args.min_coherence,
         start=args.start_date,
         end=args.end_date,
@@ -56,6 +71,19 @@ def main() -> None:
 
     if args.bucket:
         upload_file_to_s3(product_file, args.bucket, args.bucket_prefix)
+
+    if args.publish_bucket:
+        if args.publish_prefix:
+            prefix = args.publish_prefix
+            s3_name = str(product_file)
+        else:
+            if args.start_date is None or args.end_date is None:
+                prefix = 'multiburst_ts/'
+                s3_name = str(product_file)
+            else:
+                prefix = f'multiburst_ts/{str(product_file).split(".zip")[0]}'
+                s3_name = f'{args.start_date.replace("-", "")}_{args.end_date.replace("-", "")}.zip'
+        upload_file_to_s3_with_publish_access_keys(product_file, args.publish_bucket, prefix, s3_name)
 
 
 if __name__ == '__main__':